requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
tensorflow==2.20.0
numpy==2.3.4
pillow==12.0.0
python-multipart==0.0.20